Object interaction script and edits to melee script
I created the object interaction script that checks if an object is of the "interactable layer" and then makes a call to a given text canvas to display the interaction text button. I also added the ability to change the interact key and range, as well as a visual to tell where and when the raycast is making contact.