Small changes to object interaction  and melee script
Added the ability for the camera you want to use being changed in the inspector. Added the ability for the weapons box collider to enable/disable depending on if you are attacking. Still having issues with it enabling and disabling  without the function call being called with the attack cooldown. I'll continue to tinker with it.