Screenshot Tool Added
In my playtest scene under "ZaneTesting" you should be able to move the camera in front of an object and press "T" to take a screenshot. Sadly, the screenshot goes into the project folder in the file explorer and not in the asset folder. Nevertheless, it does take screenshots which are needed for the inventory icons.