Stamina Item Pickup
Created a manager that controls the spawning and updating logic of the items spawned. Will also add a tag on the item on instantiation if it is not already tagged. Most of the logic is there, I just need to add a reference to the Save Manager; I wasn't able to find it in the project.