Edit 3
Should now be able to have spawn-points added to the list in the inspector instead of searching through children.

I was still getting the error we saw when there was no SerializeField designation in the script. At this point I think it has to do with exposing the list itself to the inspector. But that doesn't make any sense...IDK.